personal_finance/requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coolname==2.2.0
croniter==4.0.0
cryptography==43.0.3
dateparser==1.2.0
Deprecated==1.2.14
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.4
fsspec==2024.10.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
oauthlib==3.2.2
opentelemetry-api==1.28.1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
prefect==3.1.1
prometheus_client==0.21.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
wrapt==1.16.0
zipp==3.20.2
